# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SourceUrlInput {
  /** @minLength 8 */
  url: string;
}

export interface ClipInput {
  /** @minLength 8 */
  url: string;
  /** @minimum 0 */
  startSeconds: number;
  /** @exclusiveMinimum 0 */
  endSeconds: number;
  /**
     * @maxLength 120
     * @nullable
     */
  title?: string | null;
}

export type VideoInfoPlatform = typeof VideoInfoPlatform[keyof typeof VideoInfoPlatform];


export const VideoInfoPlatform = {
  youtube: 'youtube',
  facebook: 'facebook',
} as const;

export interface VideoInfo {
  url: string;
  platform: VideoInfoPlatform;
  title: string;
  durationSeconds: number;
  /** @nullable */
  thumbnailUrl: string | null;
  /** @nullable */
  previewUrl: string | null;
  filename: string;
}

export interface VideoError {
  error: string;
  code?: string;
}

export type DownloadVideoParams = {
url: string;
};