# lib/api-zod/src/generated/types/clipInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ClipInput {
  /** @minLength 8 */
  url: string;
  /** @minimum 0 */
  startSeconds: number;
  /** @exclusiveMinimum 0 */
  endSeconds: number;
  /**
     * @maxLength 120
     * @nullable
     */
  title?: string | null;
}