# lib/api-zod/src/generated/types/videoInfo.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VideoInfoPlatform } from './videoInfoPlatform';

export interface VideoInfo {
  url: string;
  platform: VideoInfoPlatform;
  title: string;
  durationSeconds: number;
  /** @nullable */
  thumbnailUrl: string | null;
  /** @nullable */
  previewUrl: string | null;
  filename: string;
}